# crates/crucible-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crucible web API",
    "description": "The HTTP API that `cru web` serves. This document is generated from the axum router; do not edit it by hand.",
    "version": "0.30.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "summary": "`GET /api/config` — the effective config, its per-leaf provenance, and the\ntwo fields this route has always served.",
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not report its config"
          }
        }
      },
      "post": {
        "summary": "`POST /api/config` — save values as the user's durable preference.",
        "description": "The body and the answer are `config.save`'s own: `{values}` in,\n`{ok, refused, rejected}` out. A refusal rides in the answer rather than in\nan HTTP error because refusal is per leaf — the siblings the user changed\nin the same click did save — and because the caller needs the file and the\nline the daemon named, which an error status cannot carry.",
        "operationId": "save_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigSaveReply"
                }
              }
            }
          },
          "502": {
            "description": "The daemon could not save the values"
          }
        }
      }
    },
...
    "/api/plugins/publications": {
      "get": {
        "summary": "`GET /api/plugins/publications` — what plugins published about themselves.",
        "description": "The two levels are read; the values are passed through verbatim. Nothing\nhere interprets one, which is the point: the frontend used to learn what\nisolation a box offered by having the server match on the shape of the `oci`\nplugin's config, so one plugin's schema lived in the rendering layer and a\nsecond isolating plugin would not have appeared at all.",
        "operationId": "list_publications",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Narrow to one contribution kind.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-crucible-plugin",
            "in": "header",
            "description": "Who is asking: `app`, or the plugin being drawn for. A request without it is refused.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginPublicationsResponse"
                }
              }
            }
          },
          "403": {
            "description": "No caller identity was sent"
          },
          "502": {
            "description": "The daemon could not read the publications, or answered a shape this route cannot read"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigResponse": {
        "type": "object",
        "description": "What `GET /api/config` answers.\n\nThree of its fields stay open objects. The effective config, the origin\nrows and the control tree are the daemon's vocabulary, declared in Lua and\nin the store; a fixed shape here would make this layer a second owner of\nthem, and a key a newer daemon adds would not reach the browser at all.",
        "required": [
          "kiln_path",
          "remote_shell",
          "config",
          "origins",
          "controls"
        ],
        "properties": {
          "kiln_path": {
            "type": "string"
          },
          "remote_shell": {
            "type": "boolean",
            "description": "Non-loopback terminal/shell access is active (opt-in + API key) —\nthe terminal panel connects from LAN clients only when this is true."
          },
          "config": {
            "type": "object",
            "description": "The daemon's effective config, whole and unrewritten."
          },
          "config_root": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where `init.lua` and `settings.json` live, so a lock can offer a jump\nto the file that holds a key. Absent when the daemon booted from a\nconfig value rather than a file."
          },
          "origins": {
            "type": "object",
            "description": "One row per recorded leaf: `{key, value, source, file?, line?}`, as\n`config.origin` gives them. The flat shape is what a settings control\nrenders a lock from; the effective config's own `provenance` map is\nthe same fact in the store's enum shape, and serving both would be two\nspellings of one answer."
          },
          "controls": {
            "type": "object",
            "description": "The declared control tree the settings UI renders, as\n`config.controls` gives it: `{options, read_only}`. Served here rather\nthan from a second endpoint because a control and the value it shows\nare one screen, and two fetches could disagree about which keys exist."
          }
        }
      },
      "ConfigSaveReply": {
        "type": "object",
        "description": "What `config.save` answers.\n\nA refusal rides in the answer rather than in an error: refusal is per leaf,\nthe siblings the caller changed in the same call did save, and `refused`\ncarries the file and the line a human's config holds the key on.",
        "required": [
          "ok",
          "refused",
          "rejected"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Whether every leaf the caller sent reached the `Settings` layer."
          },
          "refused": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PinnedLeaf"
            },
            "description": "The leaves a pin refused, each with the source that holds it."
          },
          "rejected": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The top-level keys that name where the daemon acts, which no save may\nwrite. They are dropped rather than refused, so they are reported apart\nfrom `refused`."
          }
        }
      },
...
      "PinnedLeaf": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SourceOrigin",
            "description": "The source that holds the leaf, with the file and line when it has\nthem, so the caller can offer a jump to that line."
          },
          {
            "type": "object",
            "required": [
              "key"
            ],
            "properties": {
              "key": {
                "type": "string",
                "description": "The dot-joined leaf path the caller asked to save."
              }
            }
          }
        ],
        "description": "One leaf `config.save` refuses, and what pins it."
      },
...
      "PluginPublicationsResponse": {
        "type": "object",
        "description": "What `GET /api/plugins/publications` answers.",
        "required": [
          "publications"
        ],
        "properties": {
          "publications": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
      "SaveRequest": {
        "type": "object",
        "description": "The values one save carries, in the shape `config.save` takes.\n\nThe keys are config leaf paths and the values are whatever those leaves\nhold, so the map stays open: which leaves exist is the daemon's answer.",
        "required": [
          "values"
        ],
        "properties": {
          "values": {
            "type": "object"
          }
        }
      },
...
      "SourceOrigin": {
        "type": "object",
        "description": "Where one leaf came from, in the shape the wire uses.\n\nOne projection serves two callers: `config.origin` answers with it, and a\n`config.save` refusal names the pin with it. Two projections would let the\nrefusal name a file the origin does not.\n`Deserialize` as well as `Serialize`: `config.save`'s refusal travels back\nover the RPC into a typed reply, and a `&'static str` cannot be read from a\ndocument. The value is still [`ConfigSource::short`]'s, written once.",
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string",
            "description": "The one-word source name, as [`ConfigSource::short`] gives it."
          },
          "file": {
            "type": [
              "string",
              "null"
            ],
            "description": "The file the source names, when it names one."
          },
          "line": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The line inside `file`, when the source recorded one.",
            "minimum": 0
          }
        }
      },
...
    }
  }
}